Reorg scan test files to be consistent with other functional tests.
Also, enabled scan tests. Don't know how they got turned off.
