+++ test/functional/CMakeLists.txt
@@ -8,3 +8,5 @@
 add_subdirectory(atomic)
 
 add_subdirectory(forall)
+
+add_subdirectory(scan)
