+++ test/functional/scan/tests/test-scan-exclusive.hpp
@@ -10,21 +10,20 @@
 
 #include <numeric>
 
-#include "test-scan.hpp"
+#include "test-scan-utils.hpp"
 
-template <typename OP>
-::testing::AssertionResult check_inclusive(
-  const typename OP::result_type* actual,
-  const typename OP::result_type* original,
-  int N)
+template <typename OP, typename T>
+::testing::AssertionResult check_exclusive(const T* actual,
+                                           const T* original,
+                                           int N,
+                                           T init = OP::identity())
 {
-  typename OP::result_type init = OP::identity();
   for (int i = 0; i < N; ++i) {
-    init = OP()(init, *original);
     if (*actual != init) {
       return ::testing::AssertionFailure()
              << *actual << " != " << init << " (at index " << i << ")";
     }
+    init = OP()(init, *original);
     ++actual;
     ++original;
   }
@@ -32,7 +31,9 @@ template <typename OP>
 }
 
 template <typename EXEC_POLICY, typename WORKING_RES, typename OP_TYPE>
-void ScanInclusiveFunctionalTest(int N)
+void ScanExclusiveFunctionalTest(int N,
+                                 typename OP_TYPE::result_type offset = 
+                                 OP_TYPE::identity())
 {
   using T = typename OP_TYPE::result_type;
 
@@ -52,14 +53,15 @@ void ScanInclusiveFunctionalTest(int N)
 
   working_res.memcpy(work_in, host_in, sizeof(T) * N);
 
-  RAJA::inclusive_scan<EXEC_POLICY>(work_in,
+  RAJA::exclusive_scan<EXEC_POLICY>(work_in,
                                     work_in + N,
                                     work_out,
-                                    OP_TYPE{});
+                                    OP_TYPE{},
+                                    offset);
 
   working_res.memcpy(host_out, work_out, sizeof(T) * N);
 
-  ASSERT_TRUE(check_inclusive<OP_TYPE>(host_out, host_in, N));
+  ASSERT_TRUE(check_exclusive<OP_TYPE>(host_out, host_in, N, offset));
 
   deallocScanTestData(working_res,
                       work_in, work_out,             
@@ -67,7 +69,9 @@ void ScanInclusiveFunctionalTest(int N)
 }
 
 template <typename EXEC_POLICY, typename WORKING_RES, typename OP_TYPE>
-void ScanInclusiveInplaceFunctionalTest(int N)
+void ScanExclusiveInplaceFunctionalTest(int N,
+                                        typename OP_TYPE::result_type offset =
+                                        OP_TYPE::identity())
 {
   using T = typename OP_TYPE::result_type;
 
@@ -87,49 +91,74 @@ void ScanInclusiveInplaceFunctionalTest(int N)
 
   working_res.memcpy(work_in, host_in, sizeof(T) * N);
 
-  RAJA::inclusive_scan_inplace<EXEC_POLICY>(work_in,
+  RAJA::exclusive_scan_inplace<EXEC_POLICY>(work_in,
                                             work_in + N,
-                                            OP_TYPE{});
+                                            OP_TYPE{},
+                                            offset);
 
   working_res.memcpy(host_out, work_in, sizeof(T) * N);
 
-  ASSERT_TRUE(check_inclusive<OP_TYPE>(host_out, host_in, N));
+  ASSERT_TRUE(check_exclusive<OP_TYPE>(host_out, host_in, N, offset));
 
   deallocScanTestData(working_res,
                       work_in, work_out,
                       host_in, host_out);
 }
 
-TYPED_TEST_P(ScanFunctionalTest, ScanInclusive)
+TYPED_TEST_P(ScanFunctionalTest, ScanExclusive)
 {
-  using EXEC_POLICY      = typename at<TypeParam, num<0>>::type;
-  using WORKING_RESOURCE = typename at<TypeParam, num<1>>::type;
-  using OP_TYPE          = typename at<TypeParam, num<2>>::type;
+  using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<0>>::type;
+  using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
+  using OP_TYPE          = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  ScanInclusiveFunctionalTest<EXEC_POLICY, 
+  ScanExclusiveFunctionalTest<EXEC_POLICY, 
                               WORKING_RESOURCE, 
                               OP_TYPE>(357);
-  ScanInclusiveFunctionalTest<EXEC_POLICY, 
+  ScanExclusiveFunctionalTest<EXEC_POLICY, 
                               WORKING_RESOURCE, 
                               OP_TYPE>(32000);
+
+  //
+  // Perform some non-identity offset tests
+  // 
+  using T = typename OP_TYPE::result_type;
+
+  ScanExclusiveFunctionalTest<EXEC_POLICY, 
+                              WORKING_RESOURCE, 
+                              OP_TYPE>(357, T(15));
+  ScanExclusiveFunctionalTest<EXEC_POLICY, 
+                              WORKING_RESOURCE, 
+                              OP_TYPE>(32000, T(2));
 }
 
 TYPED_TEST_P(ScanFunctionalTest, ScanInclusiveInplace)
 {
-  using EXEC_POLICY      = typename at<TypeParam, num<0>>::type;
-  using WORKING_RESOURCE = typename at<TypeParam, num<1>>::type;
-  using OP_TYPE          = typename at<TypeParam, num<2>>::type;
+  using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<0>>::type;
+  using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
+  using OP_TYPE          = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  ScanInclusiveInplaceFunctionalTest<EXEC_POLICY, 
+  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY, 
                                      WORKING_RESOURCE,
                                      OP_TYPE>(357);
-  ScanInclusiveInplaceFunctionalTest<EXEC_POLICY,
+  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
                                      WORKING_RESOURCE,
                                      OP_TYPE>(32000);
+
+  //
+  // Perform some non-identity offset tests
+  //
+  using T = typename OP_TYPE::result_type;
+
+  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
+                                     WORKING_RESOURCE,
+                                     OP_TYPE>(357, T(15));
+  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
+                                     WORKING_RESOURCE,
+                                     OP_TYPE>(32000, T(2));
 }
 
 REGISTER_TYPED_TEST_SUITE_P(ScanFunctionalTest, 
-                            ScanInclusive,
+                            ScanExclusive,
                             ScanInclusiveInplace);
 
 #endif // __TEST_SCAN_INCLUSIVE_HPP__
