+++ test/functional/scan/tests/test-scan-inclusive.hpp
@@ -10,20 +10,21 @@
 
 #include <numeric>
 
-#include "test-scan.hpp"
+#include "test-scan-utils.hpp"
 
-template <typename OP, typename T>
-::testing::AssertionResult check_exclusive(const T* actual,
-                                           const T* original,
-                                           int N,
-                                           T init = OP::identity())
+template <typename OP>
+::testing::AssertionResult check_inclusive(
+  const typename OP::result_type* actual,
+  const typename OP::result_type* original,
+  int N)
 {
+  typename OP::result_type init = OP::identity();
   for (int i = 0; i < N; ++i) {
+    init = OP()(init, *original);
     if (*actual != init) {
       return ::testing::AssertionFailure()
              << *actual << " != " << init << " (at index " << i << ")";
     }
-    init = OP()(init, *original);
     ++actual;
     ++original;
   }
@@ -31,9 +32,7 @@ template <typename OP, typename T>
 }
 
 template <typename EXEC_POLICY, typename WORKING_RES, typename OP_TYPE>
-void ScanExclusiveFunctionalTest(int N,
-                                 typename OP_TYPE::result_type offset = 
-                                 OP_TYPE::identity())
+void ScanInclusiveFunctionalTest(int N)
 {
   using T = typename OP_TYPE::result_type;
 
@@ -53,15 +52,14 @@ void ScanExclusiveFunctionalTest(int N,
 
   working_res.memcpy(work_in, host_in, sizeof(T) * N);
 
-  RAJA::exclusive_scan<EXEC_POLICY>(work_in,
+  RAJA::inclusive_scan<EXEC_POLICY>(work_in,
                                     work_in + N,
                                     work_out,
-                                    OP_TYPE{},
-                                    offset);
+                                    OP_TYPE{});
 
   working_res.memcpy(host_out, work_out, sizeof(T) * N);
 
-  ASSERT_TRUE(check_exclusive<OP_TYPE>(host_out, host_in, N, offset));
+  ASSERT_TRUE(check_inclusive<OP_TYPE>(host_out, host_in, N));
 
   deallocScanTestData(working_res,
                       work_in, work_out,             
@@ -69,9 +67,7 @@ void ScanExclusiveFunctionalTest(int N,
 }
 
 template <typename EXEC_POLICY, typename WORKING_RES, typename OP_TYPE>
-void ScanExclusiveInplaceFunctionalTest(int N,
-                                        typename OP_TYPE::result_type offset =
-                                        OP_TYPE::identity())
+void ScanInclusiveInplaceFunctionalTest(int N)
 {
   using T = typename OP_TYPE::result_type;
 
@@ -91,74 +87,49 @@ void ScanExclusiveInplaceFunctionalTest(int N,
 
   working_res.memcpy(work_in, host_in, sizeof(T) * N);
 
-  RAJA::exclusive_scan_inplace<EXEC_POLICY>(work_in,
+  RAJA::inclusive_scan_inplace<EXEC_POLICY>(work_in,
                                             work_in + N,
-                                            OP_TYPE{},
-                                            offset);
+                                            OP_TYPE{});
 
   working_res.memcpy(host_out, work_in, sizeof(T) * N);
 
-  ASSERT_TRUE(check_exclusive<OP_TYPE>(host_out, host_in, N, offset));
+  ASSERT_TRUE(check_inclusive<OP_TYPE>(host_out, host_in, N));
 
   deallocScanTestData(working_res,
                       work_in, work_out,
                       host_in, host_out);
 }
 
-TYPED_TEST_P(ScanFunctionalTest, ScanExclusive)
+TYPED_TEST_P(ScanFunctionalTest, ScanInclusive)
 {
-  using EXEC_POLICY      = typename at<TypeParam, num<0>>::type;
-  using WORKING_RESOURCE = typename at<TypeParam, num<1>>::type;
-  using OP_TYPE          = typename at<TypeParam, num<2>>::type;
+  using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<0>>::type;
+  using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
+  using OP_TYPE          = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  ScanExclusiveFunctionalTest<EXEC_POLICY, 
+  ScanInclusiveFunctionalTest<EXEC_POLICY, 
                               WORKING_RESOURCE, 
                               OP_TYPE>(357);
-  ScanExclusiveFunctionalTest<EXEC_POLICY, 
+  ScanInclusiveFunctionalTest<EXEC_POLICY, 
                               WORKING_RESOURCE, 
                               OP_TYPE>(32000);
-
-  //
-  // Perform some non-identity offset tests
-  // 
-  using T = typename OP_TYPE::result_type;
-
-  ScanExclusiveFunctionalTest<EXEC_POLICY, 
-                              WORKING_RESOURCE, 
-                              OP_TYPE>(357, T(15));
-  ScanExclusiveFunctionalTest<EXEC_POLICY, 
-                              WORKING_RESOURCE, 
-                              OP_TYPE>(32000, T(2));
 }
 
 TYPED_TEST_P(ScanFunctionalTest, ScanInclusiveInplace)
 {
-  using EXEC_POLICY      = typename at<TypeParam, num<0>>::type;
-  using WORKING_RESOURCE = typename at<TypeParam, num<1>>::type;
-  using OP_TYPE          = typename at<TypeParam, num<2>>::type;
+  using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<0>>::type;
+  using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
+  using OP_TYPE          = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY, 
+  ScanInclusiveInplaceFunctionalTest<EXEC_POLICY, 
                                      WORKING_RESOURCE,
                                      OP_TYPE>(357);
-  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
+  ScanInclusiveInplaceFunctionalTest<EXEC_POLICY,
                                      WORKING_RESOURCE,
                                      OP_TYPE>(32000);
-
-  //
-  // Perform some non-identity offset tests
-  //
-  using T = typename OP_TYPE::result_type;
-
-  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
-                                     WORKING_RESOURCE,
-                                     OP_TYPE>(357, T(15));
-  ScanExclusiveInplaceFunctionalTest<EXEC_POLICY,
-                                     WORKING_RESOURCE,
-                                     OP_TYPE>(32000, T(2));
 }
 
 REGISTER_TYPED_TEST_SUITE_P(ScanFunctionalTest, 
-                            ScanExclusive,
+                            ScanInclusive,
                             ScanInclusiveInplace);
 
 #endif // __TEST_SCAN_INCLUSIVE_HPP__
